refactor(client/): Move `FormHandles.ts` to utils/collection-form
